libs/python/starlink-client/requirements.txt
grpcio
grpcio-status
proto-plus
protobuf
requests
httpx
pydantic